requirements.txt
mysql_replication==0.18
Flask==1.0.2
Flask_API==1.0
Flask_SQLAlchemy==2.3.2
mysql_connector_repackaged==0.3.1
